authorize the controller

+++ app/Http/Controllers/ExpenseController.php
@@ -12,6 +12,8 @@ class ExpenseController extends Controller
 {
     public function store(ExpenseStoreRequest $request)
     {
+        $this->authorize('create', Expense::class);
+
         $expense = new Expense;
         $expense->description = $request->description;
         $expense->date = $request->date;
@@ -27,6 +29,8 @@ class ExpenseController extends Controller
 
     public function index()
     {
+        $this->authorize('viewAny', Expense::class);
+
         $expenses = Expense::where('user_id', Auth::id())->get();
 
         return view('home', compact('expenses'));
@@ -34,19 +38,16 @@ class ExpenseController extends Controller
 
     public function edit(Expense $expense)
     {
+        $this->authorize('update', $expense);
+
         // Retrieve the expense for editing
         // You can pass the $expense object to the edit view
         return view('expenses.edit', compact('expense'));
     }
 
-    public function update(Request $request, Expense $expense)
+    public function update(ExpenseStoreRequest $request, Expense $expense)
     {
-        // Validate the updated expense data
-        $validated = $request->validate([
-            'description' => 'required|max:191',
-            'date' => 'required|date|before_or_equal:today',
-            'amount' => 'required|numeric|min:0',
-        ]);
+        $this->authorize('update', $expense);
 
         // Update the expense with the new data
         $expense->description = $request->description;
@@ -60,7 +61,8 @@ class ExpenseController extends Controller
 
     public function destroy(Expense $expense)
     {
-        // Delete the expense
+        $this->authorize('delete', $expense);
+
         $expense->delete();
 
         // Redirect back to the expenses index page or any other appropriate page
